Improve command usage format consistency in 'balena help' output
Change-type: patch
