+++ tests/commands/help.spec.ts
@@ -58,9 +58,9 @@ Additional commands:
     device rm <uuid>                      remove a device
     device shutdown <uuid>                shutdown a device
     devices supported                     list all supported devices
-    env add name [value]                  add an environment or config variable to an application or device
-    env rename id value                   change the value of an environment variable for an app or device
-    env rm id                             remove an environment variable from an application or device
+    env add <name> [value]                add an environment or config variable to an application or device
+    env rename <id> <value>               change the value of an environment variable for an app or device
+    env rm <id>                           remove an environment variable from an application or device
     envs                                  list the environment or config variables of an app or device
     key <id>                              list a single ssh key
     key add <name> [path]                 add a SSH key to balena
